Resolve build errors, update dependencies, add flake to support dev and build. (#2)
* Fix vendorHash, set go version at 1.24.0

The latest stable version of go is 1.24.

* Remove usage of ClusterName.

This is an unsupported and no longer existing field in the podSpec. It
was never fully implemented and should not be used.

* Update virtual-kubelet to be compatible with k8s 1.31+

* Update flake, remove darwin targets.

This requires the systemd libs, which afaik are only available on
linux systems. Cross compilation should still be possible if it's
configured via nix.